refactor: rename TravelCard to RideCard
